requirements.txt
Flask
flask-login
flask-sqlalchemy
requests
playwright
pytesseract
pdf2image
filetype
Pillow
werkzeug
python-dotenv
gunicorn